streaming_and_tracing_concept/pyproject.toml
[project]
name = "streaming-and-tracing-concept"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.96.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
